Fix bug when king can't be checkmated
gameplay.rb:
- Add break statement to exit gameloop when #checkmated? returns true

checkmate_logic.rb changes:
- Fix bug that prevents checkmate stores king as own_chesspieces by
removing king from own_chesspieces
- Fix #king_escape? when it returned the opposite boolean value
when the king can't get out of check

Known bug: Not putting king in checkmate when it's suppose to be checkmate

After #chesspiece_to_protect_king? generate a chess piece move it
passes to #chess_path_intercept? to check if the generate moves can
block an opponent chess piece's path. However, #chesspiece_to_protect_king?
generate all path including path in the opppsite direction of the king.
That means the king's own piece will check an opponent path that is
in an opposite direction from the king. This cause a bug when the king
doesn't enter checkmate when the king is suppose to be in one.