Add code that prevents pawn from moving forward when blocked
Known bug
- If a player tries to move the pawn forward when it's blocked
it won't move the pawn but still go to the next turn
- when king is surrounded by pieces that can't protect him, king
can't go into checkmate when an opponent traps the king. This is
because the king's moves still generate when the king's own piece is
on that space.
- Moving backwards with a pawn in the game cause the game to
go to the next turn when it's an invalid move.