Fix bug where pawn can't protect king against opponent knight
Added #opponent_chesspiece_nearby_own_piece to #chesspiece_to_protect_king?
to fix a bug where a pawn can't protect the king against an opponent's
knight, when it's next to the knight.

Known bug:
- Pawn still able to protect king when knight is in front of it.
The Pawn should not count as a piece that can protect the king when
king is in check. Only count the pawn if it's to the side of the opponent.